Major update and fixes
Added:
Bouncy, Speedy, Sticky and Clear ink types.
Tweaked the movement of the MC.
Created a UI to display the kind of ink used,

TO DO:
Fix infinite jumping.
Fix Speedy not applying in the air (because the speed had to be locked so it wouldn't increase until infinity)
Fix Speedy and Sticky ink colours not working on the Ground.